requirements.txt
langchain==0.0.352
pandas==2.1.4
streamlit==1.29.0
matplotlib==3.8.2
Pillow==10.1.0
chromadb==0.4.21
openai==1.6.1
tiktoken==0.5.2
dotenv==0.0.5
tabulate==0.9.0